pyproject.toml
[project]
name = "p3-ppo-sac"
version = "0.1.0"
description = "P3 DRL PPO and SAC on Walker2D with pixels observations"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gymnasium[classic-control,mujoco]>=0.29",
    "matplotlib>=3.10.8",
    "mujoco>=3.0.0",
    "numpy>=2.4.3",
    "opencv-python>=4.13.0.92",
    "pyyaml>=6.0.3",
    "tensorboard>=2.20.0",
    "torch>=2.0",
]

[project.scripts]
p3 = "main:main"
